schema outline for postgres
